+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Added wildcards
 * Option to use → instead of first /
 * Word boundaries can be included in groups
+* Added rule exceptions
 
 ## 1.0 (2021-03-09)
 * Created changelog